created user interface
I created a user interface via Sketch 3 on the Mac, but not sure if it will display correctly through these links -- you may have to have sketch for it to work, and Sketch is exclusive to Macs. Else, it's a nice-looking user interface for the login page and if I can find a way to extract the files then they can be imported into the master code. I will send you guys a picture via text.